Nginx support
Add a dedicated derivation providing the nginx configuration, in the
same spirit of Apache's httpd reading a `.htaccess`. It's up to the
callee to actually use the configuration file, nothing magic happen.

A more extensive option could be to provide a NixOS module serving the
website to make it a bit more consistent but that's quite some work. For
now I'll stick to it being specified and maintained in
[ptitfred/personal-infrastructure](https://github.com/ptitfred/personal-infrastructure).

It includes an end-to-end test of a VM with nginx using the extra
configuration to make sure it's usable.

At some point some specific tests could be written, notably to ensure
redirections for old URLs are working and not lost along the way.