Deployable tool to take screenshots
In a perfect world I would take screenshots at build time and include
them in the website before deployment. For now it doens't work, mainly
because of chromium not reacting well to nix sandboxing.

The only reasonable fallback I can see is to hook the screenshots script
at deployment. It will probably be done via a one-shot systemd unit
triggered by fresh deployment. It requires a bit more configuration to
let the website know where to look assets from. A dedicated location
in the virtualhost from a local directory could do the trick.